Refactor tfvars structure
