--- modules/vhub/.header.md
@@ -1,78 +1,0 @@
-# Palo Alto Networks vHub Module for Azure
-
-A Terraform module for deploying a Virtual Hub (vHub) and its components required for the firewalls in Azure.
-
-## Usage
-
-This module is designed to work in several *modes* depending on which variables or flags are set. Most common usage scenarios are:
-
-- create all - creates a  vHub, Connections and Route Tables. In this mode a vHub is created. The Connections and Route Tables are assigned to the vHub:
-
-  ```hcl
-    virtual_hubs = {
-      "virtual_hub" = {
-        name           = "virtual_hub"
-        address_prefix = "11.0.0.0/24"
-        connections = {
-          "panorama-to-hub" = { # TODO: Specify your existing panorama vnet!
-            name                       = "panorama-to-hub"
-            connection_type            = "Vnet"
-            remote_virtual_network_key = "panorama"
-          }
-          "app1-to-hub" = {
-            name                       = "app1-to-hub"
-            connection_type            = "Vnet"
-            remote_virtual_network_key = "app1"
-          }
-          "app2-to-hub" = {
-            name                       = "app2-to-hub"
-            connection_type            = "Vnet"
-            remote_virtual_network_key = "app2"
-          }
-        }
-        route_tables = {
-          "default" = {
-            name   = "default-rt"
-            labels = ["default-rt"]
-          }
-        }
-
-      }
-    }
-  ```
-
-- sources a vHub but creates Connections and Route Tables:
-
-  ```hcl
-    virtual_hubs = {
-      "virtual_hub" = {
-        create              = false
-        name                = "virtual_hub"
-        resource_group_name = "virtual_wan_rg"
-        connections = {
-          "panorama-to-hub" = { # TODO: Specify your existing panorama vnet!
-            name                       = "panorama-to-hub"
-            connection_type            = "Vnet"
-            remote_virtual_network_key = "panorama"
-          }
-          "app1-to-hub" = {
-            name                       = "app1-to-hub"
-            connection_type            = "Vnet"
-            remote_virtual_network_key = "app1"
-          }
-          "app2-to-hub" = {
-            name                       = "app2-to-hub"
-            connection_type            = "Vnet"
-            remote_virtual_network_key = "app2"
-          }
-        }
-        route_tables = {
-          "default" = {
-            name   = "default-rt"
-            labels = ["default-rt"]
-          }
-        }
-      }
-    }
-  ```
-
