--- modules/vhub/main_test.go
@@ -1,11 +1,0 @@
-package vhub
-
-import (
-	"testing"
-
-	"github.com/PaloAltoNetworks/terraform-modules-swfw-tests-skeleton/pkg/testskeleton"
-)
-
-func TestValidate(t *testing.T) {
-	testskeleton.ValidateCode(t, nil)
-}
