--- modules/vhub_routing/.header.md
@@ -1,48 +1,0 @@
-# Palo Alto Networks Virtual HUB Routing Module for Azure
-
-A Terraform module for managing routing in an Azure Virtual HUB.
-
-## Usage
-
-In order to use the `vhub_routing` module, you need to deploy the `vwan` and `vhub` module as a prerequisite.
-Then you can use the example below to call the module and configure routing in the Virtual Hub:
-
-```hcl
-module "vhub_routing" {
-  source = "../../modules/vhub_routing"
-
-  for_each = var.virtual_wan.virtual_hubs
-
-  routing_intent = merge(each.value.routing_intent, {
-    routing_policy = [
-      for policy in each.value.routing_intent.routing_policy : merge(policy, {
-        next_hop_id = module.cloudngfw[policy.next_hop_key].palo_alto_virtual_network_appliance_id
-      })
-    ]
-  })
-  virtual_hub_id = module.virtual_hub[each.key].virtual_hub_id
-}
-```
-
-Below there are provided sample values for `virtual_hub_routing` map:
-
-```hcl
-virtual_hub_routing = {
-  routing_intent = {
-    routing_intent_name = "routing_intent"
-    routing_policy = [
-      {
-        routing_policy_name = "PrivateTraffic"
-        destinations        = ["PrivateTraffic"]
-        next_hop_key        = "cloudngfw"
-      },
-      {
-        routing_policy_name = "Internet"
-        destinations        = ["Internet"]
-        next_hop_key        = "cloudngfw"
-      }
-    ]
-  }
-}
-```
-
