--- modules/vhub_routing/versions.tf
@@ -1,9 +1,0 @@
-terraform {
-  required_version = ">= 1.5, < 2.0"
-  required_providers {
-    azurerm = {
-      source  = "hashicorp/azurerm"
-      version = "~> 4.0"
-    }
-  }
-}
