+++ modules/vwan/.header.md
@@ -1,7 +1,169 @@
 # Palo Alto Networks vWAN Module for Azure
 
-A Terraform module for deploying a Virtual WAN (vWAN) required for the firewalls in Azure.
+This Terraform module manages core Azure Virtual WAN networking components. It enables the creation and configuration of a Virtual WAN, one or more Virtual Hubs, Hub Route Tables, and Hub Connections (such as VNet or VPN Site connections). The module also supports deploying VPN Gateways within hubs and configuring VPN Sites with advanced IPSec settings.
+
+It allows both creating new and referencing existing Virtual WANs and Virtual Hubs, offering flexibility for greenfield and brownfield deployments.
 
 ## Usage
 
-TBD
+ -  creates a vWAN, vHUB, Connections and Routing Intent resource
+
+```hcl
+virtual_wans = {
+  "virtual_wan" = {
+    name = "virtual_wan"
+    virtual_hubs = {
+      "virtual_hub" = {
+        name           = "virtual_hub"
+        address_prefix = "11.0.0.0/24"
+        connections = {
+          "app1-to-hub" = {
+            name                       = "app1-to-hub"
+            connection_type            = "Vnet"
+            remote_virtual_network_key = "app1"
+          }
+          "app2-to-hub" = {
+            name                       = "app2-to-hub"
+            connection_type            = "Vnet"
+            remote_virtual_network_key = "app2"
+          }
+        }
+        routing_intent = {
+          routing_intent_name = "routing_intent"
+          routing_policy = [
+            {
+              routing_policy_name = "PrivateTraffic"
+              destinations        = ["PrivateTraffic"]
+              next_hop_key        = "cloudngfw"
+            },
+            {
+              routing_policy_name = "Internet"
+              destinations        = ["Internet"]
+              next_hop_key        = "cloudngfw"
+            }
+          ]
+        }
+      }
+    }
+  }
+}
+```
+
+ -  creates a vWAN, vHUB, Route Table, Routes and Connections. Associates each Connection with the appropriate Route Table based on the provided configuration
+
+```hcl
+virtual_wans = {
+  "virtual_wan" = {
+    name = "virtual_wan"
+    virtual_hubs = {
+      "virtual_hub" = {
+        name           = "virtual_hub"
+        address_prefix = "11.0.0.0/24"
+        route_tables = {
+          "route_table" = {
+            name   = "route_table"
+            labels = ["rt_label01", "rt_label02"]
+            routes = {
+              "route01" = {
+                name              = "route01"
+                destinations_type = "CIDR"
+                destinations      = ["10.0.0.0/16"]
+                next_hop_key      = "ha-ports"
+              }
+              "route02" = {
+                name              = "route02"
+                destinations_type = "CIDR"
+                destinations      = ["11.0.0.0/16"]
+                next_hop_key      = "ha-ports"
+              }
+            }
+          }
+        }
+        connections = {
+          "app1-to-hub" = {
+            name                       = "app1-to-hub"
+            connection_type            = "Vnet"
+            remote_virtual_network_key = "app1"
+            routing = {
+                associated_route_table_key =  "route_table"
+                propagated_route_table_keys = ["route_table"]
+            }
+          }
+          "app2-to-hub" = {
+            name                       = "app2-to-hub"
+            connection_type            = "Vnet"
+            remote_virtual_network_key = "app2"
+            routing = {
+                associated_route_table_key =  "route_table"
+                propagated_route_table_keys = ["route_table"]
+            }
+          }
+        }
+      }
+    }
+  }
+}
+```
+
+-  creates a vWAN, vHUB, VPN gateway in vHUB, VPN sites and Connections. Associates Connection with the appropriate VPN Site based on the provided configuration
+
+```hcl
+virtual_wans = {
+  "virtual_wan" = {
+    name = "virtual_wan"
+    virtual_hubs = {
+      "virtual_hub" = {
+        name           = "virtual_hub"
+        address_prefix = "11.0.0.0/24"
+        connections = {
+          "site-to-hub" = {
+            name            = "site-to-hub"
+            connection_type = "Site-to-Site"
+            vpn_site_key    = "site"
+            vpn_link = [
+              {
+                vpn_link_name     = "link-name"
+                vpn_site_link_key = "ISPW1"
+                bandwidth_mbps    = 50
+                ipsec_policy = {
+                  dh_group                 = "DHGroup24"
+                  ike_encryption_algorithm = "AES256"
+                  ike_integrity_algorithm  = "SHA256"
+                  encryption_algorithm     = "GCMAES128"
+                  integrity_algorithm      = "GCMAES128"
+                  pfs_group                = "PFS14"
+                  sa_data_size_kb          = 1025
+                  sa_lifetime_sec          = 301
+                }
+              }
+            ]
+          }
+        }
+        vpn_gateway = {
+          name = "vpn_gateway01"
+        }
+        vpn_sites = {
+          "site" = {
+            name          = "site"
+            address_cidrs = ["10.0.0.0/24"]
+            link = {
+              "ISPW1" = {
+                name          = "ISPW1"
+                ip_address    = "203.0.113.1"
+                provider_name = "Provider01"
+                speed_in_mbps = 100
+              }
+              "ISPW2" = {
+                name          = "ISPW2"
+                ip_address    = "205.0.113.1"
+                provider_name = "Provider02"
+                speed_in_mbps = 100
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+```
